+++ docs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 **Breaking changes:**
 
-It is no longer guaranteed that a field's value is available under `source[fieldName]` in a custom resolver. Instead, custom resolvers on non-trivial fields need to use GraphQL's default resolver to get the field value:
+It is no longer guaranteed that a field's value is available under `source[fieldName]` in a custom resolver. Instead, custom resolvers on [non-trivial fields](./warnings.md#non-trivial-fields)  need to use GraphQL's default resolver to get the field value:
 
 ```javascript
 import { defaultFieldResolver } from 'graphql'